+++ packages/calcite-components-react/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.0-next.60](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.59...@esri/calcite-components-react@3.0.0-next.60) (2024-12-12)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.59](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.58...@esri/calcite-components-react@3.0.0-next.59) (2024-12-12)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.58](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.57...@esri/calcite-components-react@3.0.0-next.58) (2024-12-12)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.57](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.56...@esri/calcite-components-react@3.0.0-next.57) (2024-12-11)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.56](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.55...@esri/calcite-components-react@3.0.0-next.56) (2024-12-11)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.55](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.54...@esri/calcite-components-react@3.0.0-next.55) (2024-12-10)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.54](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.53...@esri/calcite-components-react@3.0.0-next.54) (2024-12-10)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.53](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.52...@esri/calcite-components-react@3.0.0-next.53) (2024-12-09)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.52](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.51...@esri/calcite-components-react@3.0.0-next.52) (2024-12-06)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.51](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.50...@esri/calcite-components-react@3.0.0-next.51) (2024-12-04)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.50](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.49...@esri/calcite-components-react@3.0.0-next.50) (2024-12-04)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.49](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.48...@esri/calcite-components-react@3.0.0-next.49) (2024-12-04)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.48](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.47...@esri/calcite-components-react@3.0.0-next.48) (2024-12-04)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.47](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.46...@esri/calcite-components-react@3.0.0-next.47) (2024-12-03)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [3.0.0-next.46](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.45...@esri/calcite-components-react@3.0.0-next.46) (2024-12-03)
 
 **Note:** Version bump only for package @esri/calcite-components-react